first commit for full use of JobScheduler, many functions changed to be thread safe, main changed to use JobScheduler (except Dynamic AddEdge, minor changes/comments/bugfixes

+++ graph.h
@@ -34,9 +34,9 @@ int gFindShortestPath(pGraph, graphNode from, graphNode to, int type);
 	// Returns the length of the path found, or an appropriate error code (negative number) in case of error
 	// 'type' is the type of search, defined above
 
-int gFindShortestPath_t(pGraph g, graphNode from, graphNode to,phead *lists,pvis visited,int version);
-    //Is the thread safe version of the above function,
-    //needs 2 intlists for CC/bidirectional searches ,version will be used only for dynamic graphs
+int gFindShortestPath_t(pGraph g, graphNode from, graphNode to, phead *lists, pvis visited, unsigned long int version);
+    // Is the thread safe version of the above function,
+    // Needs 2 intlists for CC/bidirectional searches ,version will be used only for dynamic graphs
 
 int calculate_hashtable_size(pGraph g);
